:recycle: refactor: Troca funções de arquivo
