--- utils/types.ts
@@ -1,54 +1,0 @@
-/*
-import {question} from "readline-sync"
-
-function print(mensagem: string): void {
-    console.log(mensagem);
-}
-
-function getNumber(mensagem: string): number {
-    return Number(question(mensagem));
-}
-
-function getData(mensagem: string): string {
-    return question(mensagem);
-}
-
-function getNaturalNumber(numero: number): number {
-    if (numero >= 0) return numero;
-    throw new Error('Apenas valores positivos');
-}
-
-function getPositiveNumber(numero: number): number {
-    if (numero > 0) return numero;
-    throw new Error('Apenas valores positivos');
-}
-
-function getNegativeNumber(numero: number): number {
-    if (numero < 0) return numero;
-    throw new Error('Apenas valores negativos');
-}
-
-function filterList<T>(colecao: T[], criterio: (item: T) => boolean): T[] {
-    return colecao.filter(criterio);
-}
-
-function mapList<T, R>(colecao: T[], criterio: (item: T) => R): R[] {
-    return colecao.map(criterio);
-}
-
-function addList<T>(colecao: T[], item: T): T[] {
-    colecao.push(item);
-    return colecao;
-}
-
-function removeList<T>(colecao: T[], item: T): T[] {
-    return colecao.filter(i => i !== item);
-}
-
-function clear(): void {
-    getData('Pressione Enter para sair');
-    console.clear();
-}
-
-export { getData, getNumber, print, getNaturalNumber, getPositiveNumber, filterList, mapList, addList, removeList, getNegativeNumber, clear };
-*/
